Add operator function and (number) key presses
This commit adds the operator function (I am unclear as to
whether it will actually work just yet)! It also adds event listeners
for the number button presses and outputs them to the screen via
innerHTML functionality.

I need to log these numbers and I need to add the event listeners for
the various operators etc. That will come soon!