Add simple sum functionality
The calculator is far from perfect. It adds, subtracts, divides and
multiplies but, in fact, it has several bugs that mean it doesn't
properly work. However, I want to come back to this project at a later
date and refactor/rewrite my code with a greater understanding of
what I am doing.

I will therefore leave this project here for now and upload what
I have with the knowledge that I *will* come back to clean up later.